+++ packages/synapse-interface/pages/_app.tsx
@@ -46,7 +46,7 @@ import { SegmentAnalyticsProvider } from '@/contexts/SegmentAnalyticsProvider'
 
 import { Provider } from 'react-redux'
 import { store } from '@/store/store'
-import { WalletAnalyticsProvider } from '@/contexts/WalletAnalyticsProvider'
+import { UserProvider } from '@/contexts/UserProvider'
 
 import PortfolioUpdater from '@/slices/portfolio/updater'
 import TransactionsUpdater from '@/slices/transactions/updater'
@@ -146,16 +146,16 @@ const App = ({ Component, pageProps }: AppProps) => {
       <WagmiConfig config={wagmiConfig}>
         <RainbowKitProvider chains={chains} theme={darkTheme()}>
           <SynapseProvider chains={chains}>
-            <SegmentAnalyticsProvider>
-              <WalletAnalyticsProvider>
-                <Provider store={store}>
+            <Provider store={store}>
+              <SegmentAnalyticsProvider>
+                <UserProvider>
                   <Updaters />
                   <Component {...pageProps} />
                   <Analytics />
                   <CustomToaster />
-                </Provider>
-              </WalletAnalyticsProvider>
-            </SegmentAnalyticsProvider>
+                </UserProvider>
+              </SegmentAnalyticsProvider>
+            </Provider>
           </SynapseProvider>
         </RainbowKitProvider>
       </WagmiConfig>